Resolve conflict regarding OLC shortening
As has been noted in google/open-location-code#462, there's a discrepancy between this file and the OLC specification regarding the number of digits that can be removed while shortening. This PR resolves this conflict in a minimally invasive way by referencing the specification instead of claiming a different value.

This resolves the aforementioned #462, which can then be closed. Please check if version number needs a patch-level bump due to this change.